jpegxl/interop-2026/features/bitdepth-8bpc.html

<!DOCTYPE html>
<title>JPEG XL interop: 8-bit decode</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes known 8-bit JPEG XL and reports expected dimensions.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/interop-helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('../resources/3x3_srgb_lossless.jxl');
  assert_equals(img.naturalWidth, 3);
  assert_equals(img.naturalHeight, 3);
}, '8-bit JPEG XL fixture decodes to expected dimensions.');

</script>
